refactor of turn analysis into turn handlers
